+++ src/main/java/net/sf/jabref/model/groups/ExplicitGroup.java
@@ -1,33 +1,28 @@
 package net.sf.jabref.model.groups;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import java.util.Set;
-import java.util.TreeSet;
 
 import net.sf.jabref.model.entry.FieldName;
-import net.sf.jabref.model.strings.StringUtil;
-
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 
 /**
- * Select explicit bibtex entries. It is also known as static group.
- *
- * @author jzieren
+ * This group contains entries, which were manually assigned to it.
+ * Assignments are stored in the {@link FieldName#GROUPS} field.
+ * Thus, internally, we represent {@link ExplicitGroup} as a special {@link WordKeywordGroup} operating on
+ * {@link FieldName#GROUPS}.
  */
-public class ExplicitGroup extends KeywordGroup {
+public class ExplicitGroup extends WordKeywordGroup {
 
-    public static final String ID = "ExplicitGroup:";
-
+    /**
+     * Previous versions of JabRef stored the linked entries directly in the "jabref-meta" comment at the end of the
+     * file. These keys are still parsed and stored in this field.
+     */
     private final List<String> legacyEntryKeys = new ArrayList<>();
 
-    private static final Log LOGGER = LogFactory.getLog(ExplicitGroup.class);
-
-
     public ExplicitGroup(String name, GroupHierarchyType context, Character keywordSeparator) {
-        super(name, FieldName.GROUPS, name, true, false, context, keywordSeparator);
+        super(name, context, FieldName.GROUPS, name, true, keywordSeparator, true);
     }
 
     public void addLegacyEntryKey(String key) {
@@ -36,7 +31,7 @@ public class ExplicitGroup extends KeywordGroup {
 
     @Override
     public AbstractGroup deepCopy() {
-        ExplicitGroup copy = new ExplicitGroup(getName(), getContext(), keywordSeparator);
+        ExplicitGroup copy = new ExplicitGroup(getName(), getHierarchicalContext(), keywordSeparator);
         copy.legacyEntryKeys.addAll(legacyEntryKeys);
         return copy;
     }
@@ -54,46 +49,17 @@ public class ExplicitGroup extends KeywordGroup {
                 other.getHierarchicalContext()) && Objects.equals(getLegacyEntryKeys(), other.getLegacyEntryKeys());
     }
 
-    /**
-     * Returns a String representation of this group and its entries.
-     */
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append(ExplicitGroup.ID).append(
-                StringUtil.quote(getName(), AbstractGroup.SEPARATOR, AbstractGroup.QUOTE_CHAR)).
-                append(AbstractGroup.SEPARATOR).append(getContext().ordinal()).append(AbstractGroup.SEPARATOR);
-
-        // write legacy entry keys in well-defined order for CVS compatibility
-        Set<String> sortedKeys = new TreeSet<>();
-        sortedKeys.addAll(legacyEntryKeys);
-
-        for (String sortedKey : sortedKeys) {
-            sb.append(StringUtil.quote(sortedKey, AbstractGroup.SEPARATOR, AbstractGroup.QUOTE_CHAR)).append(
-                    AbstractGroup.SEPARATOR);
-        }
-        return sb.toString();
-    }
-
-    /**
-     * Remove all stored cite keys, resulting in an empty group.
-     */
     public void clearLegacyEntryKeys() {
         legacyEntryKeys.clear();
     }
 
     public List<String> getLegacyEntryKeys() {
-        return legacyEntryKeys;
-    }
-
-    @Override
-    public String getTypeId() {
-        return ExplicitGroup.ID;
+        return Collections.unmodifiableList(legacyEntryKeys);
     }
 
     @Override
     public int hashCode() {
-        return super.hashCode();
+        return Objects.hash(name, context, legacyEntryKeys);
     }
 
     @Override